Common file extensions (binary)

Common file extensions (binary)

Usage

ext_binary(flat = FALSE)

Arguments

flat

Flatten the list and return a vector?

Value

A list or vector of standard binary file extensions.

Specification

The contents of this section are shown in PDF user manual only.

Examples

ext_binary()
#> $data
#> [1] "rda"   "rds"   "RData"
#> 
#> $manfigures
#> [1] "jpg"  "jpeg" "pdf"  "png" 
#> 
#> $figure
#>  [1] "png"     "jpg"     "jpeg"    "bmp"     "gif"     "tif"     "tiff"   
#>  [8] "emf"     "svgz"    "ico"     "webp"    "eps"     "pdf"     "ppm"    
#> [15] "pgm"     "pbm"     "pnm"     "xcf"     "psd"     "graffle"
#> 
#> $src
#> [1] "o"  "so"
#> 
#> $build
#> [1] "rdb" "rdx"
#> 
#> $web
#> [1] "woff2" "woff"  "otf"   "ttf"   "eot"  
#> 
#> $mso
#>  [1] "docx" "xlsx" "pptx" "xltx" "potx" "doc"  "xls"  "ppt"  "xlsb" "xlsm"
#> 
#> $odf
#> [1] "odt" "ods" "odp" "odg" "odc" "odf" "odi" "odm" "odb"
#> 
#> $sas
#> [1] "sas7bdat" "sas7bcat" "xpt"      "xpt5"     "xpt8"    
#> 
#> $archive
#> [1] "zip" "tar" "gz"  "tgz" "bz2" "7z"  "xz" 
#> 
#> $db
#> [1] "sqlite"  "sqlite3" "dbf"     "accdb"   "mdb"    
#> 
#> $proglang
#> [1] "pyc" "jar"
#> 
#> $gettext
#> [1] "mo"
#> 
#> $geo
#> [1] "shx"  "shp"  "laz"  "sbx"  "sbn"  "nc"   "gpkg"
#> 
#> $bio
#> [1] "bam" "bai"
#> 
#> $audio
#> [1] "wav" "mp3" "mid" "ogg" "au"  "m4a"
#> 
#> $video
#> [1] "mp4"  "avi"  "mov"  "mkv"  "webm"
#> 
#> $generic
#>  [1] "bin"         "epub"        "hdf5"        "h5"          "parquet"    
#>  [6] "feather"     "msgpack"     "pickle"      "pkl"         "npy"        
#> [11] "npz"         "safetensors" "pt"          "pth"         "keras"      
#> [16] "tfrecord"    "pb"          "ckpt"        "onnx"       
#>