pkglite: Compact Package Representations

A tool, grammar, and standard to represent and exchange R package source code as text files. Converts one or more source packages to a text file and restores the package structures from the file.

See also

Author

Maintainer: Nan Xiao nan.xiao1@merck.com (ORCID)

Authors:

Yilong Zhang

Keaven Anderson keaven_anderson@merck.com

Other contributors:

Amin Shirazi [contributor]

Jeff Cheng [contributor]

Danfeng Fu [contributor]

Merck & Co., Inc., Rahway, NJ, USA and its affiliates [copyright holder]